src/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
Werkzeug==2.1.2
gunicorn==20.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.30